requirements.txt
fastapi
uvicorn
numpy
pillow
keras==3.5.0
huggingface_hub
